Split InboundMessageDecoder god file into per-template partials (#139)
InboundMessageDecoder.cs was a single 911-line file with a top-level
internal static class containing one private nested offsets class +
one decode method per inbound EntryPoint template, plus three shared
helpers (Side/OrdType/TIF mappers). Adding a new template forced
edits in the middle of the file, growing merge conflicts and review
cost on every order-entry change.

This PR converts the class to `partial` and splits the decoders into
one file per template family while keeping the on-wire behaviour and
public API surface bit-identical:

- InboundMessageDecoder.cs (130 LOC) — top doc + InboundDecodeOutcome
  enum + shared constants (PriceNull, RoutingInstructionNull,
  TimeInForceOptionalNull) + Side/OrdType/TIF mappers + classifiers.
- InboundMessageDecoder.SimpleNewOrder.cs (57)
- InboundMessageDecoder.SimpleReplace.cs (59)
- InboundMessageDecoder.SimpleCancel.cs (48)
- InboundMessageDecoder.NewOrderSingle.cs (147)
- InboundMessageDecoder.OrderCancelReplace.cs (166)
- InboundMessageDecoder.NewOrderCross.cs (228)
- InboundMessageDecoder.OrderMassAction.cs (128)

Each per-template file owns its private nested `*Offsets` class so
the offsets live next to the decoder that consumes them. The per-
template files include only the `using` directives they actually
need (System.Runtime.InteropServices, B3.Exchange.Matching, plus
B3.EntryPoint.Wire for the cross decoder which calls
EntryPointVarData).

No call sites change — every consumer still calls
`InboundMessageDecoder.TryDecodeXxx`. Gateway tests (376), Core
tests (42), Host tests (33), SyntheticTrader tests (45), and
ScenarioReplay tests (23) all pass unchanged.

This addresses the InboundMessageDecoder half of #139. The
FixpSession.cs (881 LOC) and FixpSession.Inbound.cs (1042 LOC) split
is more delicate — it touches the FIXP control plane, recv loop, and
retransmit buffer with shared mutable state — and is best done as a
sequence of micro-PRs (Outbound → Retransmit → MassAction →
Lifecycle → Inbound, mirroring the PR-121 pattern from the roadmap).
That work continues under #139 in a follow-up branch.
